Check whether /proc/sys/net/ipv6/conf/all/disable_ipv6 exists

+++ modules/tasks/network-interfaces.nix
@@ -275,7 +275,9 @@ in
                 EOF
 
                 # Disable or enable IPv6.
-                echo ${if cfg.enableIPv6 then "0" else "1"} > /proc/sys/net/ipv6/conf/all/disable_ipv6
+                if [ -e /proc/sys/net/ipv6/conf/all/disable_ipv6 ]; then
+                  echo ${if cfg.enableIPv6 then "0" else "1"} > /proc/sys/net/ipv6/conf/all/disable_ipv6
+                fi
 
                 # Set the default gateway.
                 ${optionalString (cfg.defaultGateway != "") ''
